+++ src/Modules/IMPLANT_MODULE/ImplantDesignerController.php
@@ -25,6 +25,12 @@ use stdClass;
 		accessLevel: "guest",
 		description: "Implant Designer",
 		alias: "impdesign"
+	),
+	NCA\DefineCommand(
+		command: "implantshoppinglist",
+		accessLevel: "guest",
+		description: "Implant Designer Shopping List",
+		alias: ["impshop", "implantshoplist", "impshoplist"],
 	)
 ]
 class ImplantDesignerController extends ModuleInstance {
@@ -63,8 +69,78 @@ class ImplantDesignerController extends ModuleInstance {
 		$this->db->loadCSVFile($this->moduleName, __DIR__ . "/SymbiantProfessionMatrix.csv");
 	}
 
+	/** Show a shopping list for your current implant design */
+	#[NCA\HandlesCommand("implantshoppinglist")]
+	#[NCA\Help\Group("implantdesigner")]
+	public function implantShoplistCommand(CmdContext $context): void {
+		$design = $this->getDesign($context->char->name, '@');
+		$list = new ShoppingList();
+
+		/** @var array<string,string> */
+		$lookup = $this->db->table("Cluster")
+			->asObj(Cluster::class)
+			->reduce(
+				function (array $lookup, Cluster $cluster): array {
+					$lookup[$cluster->LongName] = $cluster->OfficialName;
+					return $lookup;
+				},
+				[]
+			);
+		foreach ($this->slots as $slot) {
+			if (!property_exists($design, $slot)) {
+				continue;
+			}
+			$slotObj = $design->{$slot};
+			// Symbiants are not part of the shopping list
+			if (property_exists($slotObj, "symb") && $slotObj->symb !== null) {
+				continue;
+			}
+			$ql = empty($slotObj->ql) ? 300 : (int)$slotObj->ql;
+			$addImp = false;
+			$refined = "";
+			if ($ql > 200) {
+				$refined = "Refined ";
+			}
+			foreach (["shiny", "bright", "faded"] as $grade) {
+				if (empty($slotObj->{$grade})) {
+					continue;
+				}
+				$name = $lookup[$slotObj->{$grade}];
+				if (str_ends_with($name, "Jobe")) {
+					$name = str_replace(" Jobe", " {$refined}Jobe Cluster", $name);
+				} else {
+					$name .= " {$refined}Cluster";
+				}
+				$name .= " (QL " . $this->implantController->getClusterMinQl($ql, $grade) . "+)";
+				$listGrade = "{$grade}Clusters";
+				$list->{$listGrade} []= $name;
+				$addImp = true;
+			}
+			if ($addImp) {
+				/** @var string */
+				$longName = $this->db->table("ImplantType")
+					->where("ShortName", $slot)
+					->pluckStrings("Name")
+					->firstOrFail();
+				if ($ql > 200) {
+					$list->implants []= "{$longName} Implant Refined Empty (QL {$ql})";
+				} else {
+					$list->implants []= "Basic {$longName} Implant (QL {$ql})";
+				}
+			}
+		}
+		$blob = $this->renderShoppingList($list);
+		if (empty($blob)) {
+			$context->reply("Nothing to buy.");
+			return;
+		}
+		$msg = $this->text->makeBlob("Implant Shopping List", $blob);
+		$context->reply($msg);
+	}
+
 	/** Look at your current implant design */
 	#[NCA\HandlesCommand("implantdesigner")]
+	#[NCA\Help\Group("implantdesigner")]
 	#[NCA\Help\Epilogue(
 		"<i>Slot can be any of head, eye, ear, rarm, chest, larm, rwrist, waist, ".
 		"lwrist, rhand, legs, lhand, and feet.</i>"
@@ -77,6 +153,7 @@ class ImplantDesignerController extends ModuleInstance {
 
 	/** Remove all clusters from your current implant design */
 	#[NCA\HandlesCommand("implantdesigner")]
+	#[NCA\Help\Group("implantdesigner")]
 	public function implantdesignerClearCommand(CmdContext $context, #[NCA\Str("clear")] string $action): void {
 		$this->saveDesign($context->char->name, '@', new stdClass());
 		$msg = "Implant Designer has been cleared.";
@@ -90,6 +167,7 @@ class ImplantDesignerController extends ModuleInstance {
 
 	/** See a specific slot in your current implant design */
 	#[NCA\HandlesCommand("implantdesigner")]
+	#[NCA\Help\Group("implantdesigner")]
 	public function implantdesignerSlotCommand(CmdContext $context, PImplantSlot $slot): void {
 		$slot = $slot();
 
@@ -149,6 +227,7 @@ class ImplantDesignerController extends ModuleInstance {
 
 	/** Add a cluster to a slot in your current implant design */
 	#[NCA\HandlesCommand("implantdesigner")]
+	#[NCA\Help\Group("implantdesigner")]
 	public function implantdesignerSlotAddClusterCommand(
 		CmdContext $context,
 		PImplantSlot $slot,
@@ -228,6 +307,7 @@ class ImplantDesignerController extends ModuleInstance {
 
 	/** Set the QL for a slot in your current implant design */
 	#[NCA\HandlesCommand("implantdesigner")]
+	#[NCA\Help\Group("implantdesigner")]
 	public function implantdesignerSlotQLCommand(
 		CmdContext $context,
 		PImplantSlot $slot,
@@ -256,6 +336,7 @@ class ImplantDesignerController extends ModuleInstance {
 
 	/** Clear all clusters from a slot in your current implant design */
 	#[NCA\HandlesCommand("implantdesigner")]
+	#[NCA\Help\Group("implantdesigner")]
 	public function implantdesignerSlotClearCommand(
 		CmdContext $context,
 		PImplantSlot $slot,
@@ -279,6 +360,7 @@ class ImplantDesignerController extends ModuleInstance {
 
 	/** Show how to make a slot require a certain attribute in your current implant design */
 	#[NCA\HandlesCommand("implantdesigner")]
+	#[NCA\Help\Group("implantdesigner")]
 	public function implantdesignerSlotRequireCommand(
 		CmdContext $context,
 		PImplantSlot $slot,
@@ -319,6 +401,7 @@ class ImplantDesignerController extends ModuleInstance {
 
 	/** Show how to make a slot require a certain attribute in your current implant design */
 	#[NCA\HandlesCommand("implantdesigner")]
+	#[NCA\Help\Group("implantdesigner")]
 	public function implantdesignerSlotRequireAbilityCommand(
 		CmdContext $context,
 		PImplantSlot $slot,
@@ -411,6 +494,7 @@ class ImplantDesignerController extends ModuleInstance {
 
 	/** Show the result of your current implant design */
 	#[NCA\HandlesCommand("implantdesigner")]
+	#[NCA\Help\Group("implantdesigner")]
 	public function implantdesignerResultCommand(CmdContext $context, #[NCA\Str("result", "results")] string $action): void {
 		$blob = $this->getImplantDesignerResults($context->char->name);
 
@@ -605,6 +689,44 @@ class ImplantDesignerController extends ModuleInstance {
 					"dt" => time(),
 				],
 			);
+	}
+
+	private function renderShoppingList(ShoppingList $list): string {
+		/** @var string[] */
+		$parts = [];
+		if (!empty($list->implants)) {
+			$part = "<header2>Empty Implants<end>";
+			sort($list->implants);
+			foreach ($list->implants as $implant) {
+				$part .= "\n<tab>- {$implant}";
+			}
+			$parts []= $part;
+		}
+		if (!empty($list->shinyClusters)) {
+			$part = "<header2>Shiny Clusters<end>";
+			sort($list->shinyClusters);
+			foreach ($list->shinyClusters as $cluster) {
+				$part .= "\n<tab>- {$cluster}";
+			}
+			$parts []= $part;
+		}
+		if (!empty($list->brightClusters)) {
+			$part = "<header2>Bright Clusters<end>";
+			sort($list->brightClusters);
+			foreach ($list->brightClusters as $cluster) {
+				$part .= "\n<tab>- {$cluster}";
+			}
+			$parts []= $part;
+		}
+		if (!empty($list->fadedClusters)) {
+			$part = "<header2>Faded Clusters<end>";
+			sort($list->fadedClusters);
+			foreach ($list->fadedClusters as $cluster) {
+				$part .= "\n<tab>- {$cluster}";
+			}
+			$parts []= $part;
+		}
+		return join("\n\n", $parts);
 	}
 
 	private function getImplantDesignerBuild(string $sender): string {