+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add new setting to only give timer-based raid-points if the raid is locked.
 - Allow configuring the auction-message when no one bid on the item.
 - Add the total required nanoskill to cast all those in `!bestnanos`
+- New command `!implantshoppinglist` to get a list which implants and clusters to buy at which QL.
 
 ### Changed